requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
billiard==3.6.4.0
bitarray==2.7.6
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.0
cryptography==41.0.1
cytoolz==0.12.1
Django==4.2.2
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-environ==0.10.0
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
eth-abi==4.1.0
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.1.1
frozenlist==1.3.3
hexbytes==0.3.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
kombu==5.3.1
lru-dict==1.2.0
MarkupSafe==2.1.3
multidict==6.0.4
packaging==23.1
parsimonious==0.9.0
prompt-toolkit==3.0.38
protobuf==4.23.3
pycparser==2.21
pycryptodome==3.18.0
PyJWT==2.7.0
pyrsistent==0.19.3
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
rlp==3.0.0
six==1.16.0
sqlparse==0.4.4
toolz==0.12.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
web3==6.5.0
websockets==11.0.3
yarl==1.9.2
